% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_load_edit.R
\name{fix_factors_dlist}
\alias{fix_factors_dlist}
\title{fix_factors_dlist}
\usage{
fix_factors_dlist(dlist, classdic = "default", fr = "english",
  to = "class")
}
\arguments{
\item{dlist}{dlist object such as returned by load_directory and translate_names_dlist}

\item{classdic}{Either a path to a .csv file or a data frame object.}

\item{fr}{The column in classdic where the names are found}

\item{to}{The column to classdic where class mappings are found}
}
\value{
updated dlist
}
\description{
dlist version of fix_factors
}
