% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{translate_names}
\alias{translate_names}
\title{translate_names}
\usage{
translate_names(data, dictionary = "default", reverse = FALSE)
}
\arguments{
\item{data}{Data frame object of which the names are to be translated}

\item{dictionary}{Either a path to a .csv file or a data frame object. If no
dictionary is specified a fallback dictionary is used with a warning.}

\item{reverse}{Setting this to true everses the order of the languages,
works only if there is a 1:1 mapping, not recommended.}
}
\value{
returns the data_frame, hopefully with translated names.
}
\description{
Translates the names of a data_frame using a specified dictionary. The
dictionary can be either a data frame or a path to a .csv file. In either
case the dictionary should be made up of two colums the first with the
language to translate from and the second with the language to translate
into.
}
