% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs_load_edit.R
\name{translate_names_dlist}
\alias{translate_names_dlist}
\title{translate_names_dlist}
\usage{
translate_names_dlist(dlist, dictionary = "default", reverse = FALSE)
}
\arguments{
\item{dlist}{dlist object such as returned by load_directory}

\item{dictionary}{Either a path to a .csv file or a data frame object. If no
dictionary is specified a fallback dictionary is used with a warning.}

\item{reverse}{Setting this to true everses the order of the languages,
works only if there is a 1:1 mapping, not recommended.}
}
\value{
updated dlist
}
\description{
dlist version of translate_names
}
