% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{load_directory}
\alias{load_directory}
\title{load_directory}
\usage{
load_directory(path)
}
\arguments{
\item{path}{The path to the directory that should be read.}
}
\value{
Returns a list of data frames (tibbles) containing file contents.
}
\description{
Read .px and .csv files in a directory into a list of data frames.
}
